requirements/common.txt
# these are common files for your application if you forked tendenci
# common files for tendenci live in /requirements/tendenci.txt
# the /requirements/tendenci.txt file is overwritten on every upgrade
# https://tendenci.readthedocs.org

# Your common dependencies here.
Django>=1.8,<1.9
django-formtools>=1.0
pisa
Reportlab==3.3.0
html5lib==0.9999999
Pillow==4.2.1
anyjson>=0.2.4
django-authority>=0.4
# temp fix - the latest version of django-avatar-3.0.0 breaks profile
django-countries==3.4.1
django-form-utils>=0.1.8
django-localflavor==1.4.1
dj-pagination>=2.2.0
django-picklefield>=0.1.6
django-simple-captcha==0.5.5
django-tagging==0.4.3
django-haystack==2.4.1
django-debug-toolbar>=1.2.2
django-sql-explorer==0.9.2
feedparser>=4.1
httplib2>=0.4.0
pytz==2016.4
simplejson>=2.0.9
webcolors>=1.3.1
xlrd>=0.9.4
xlwt>=0.7.2
BeautifulSoup==3.2.1
beautifulsoup4==4.4.1
oauth2>=1.5.167
python_openid>=2.2
createsend==4.1.1
celery==3.1.23
django-celery==3.1.17
django-kombu>=0.9.4
mimeparse>=0.1.3
python-dateutil==2.5.3
pdfminer==20110515
stripe==1.62.0
pycrypto==2.6.1
# boto==2.40.0
boto3==1.5.36
django-timezones==0.2
django-ses==0.7.1
Geraldo==0.4.17
django-tastypie==0.13.3
docutils==0.12
chardet==3.0.3
psycopg2==2.6.1
gunicorn==19.5.0
django-storages==1.6.5
django-s3-folder-storage==0.5
python-memcached==1.57
Whoosh==2.7.4
simple-salesforce==0.68.1
selenium==2.53.2
raven==6.0.0
django-admin-bootstrapped==2.5.7
django-bootstrap3==7.0.1
django-app-namespace-template-loader==0.4
django-annoying
unidecode
Markdown
unicodecsv
Embedly>=0.5.0
# dependencies for helpdesk
django-bootstrap-form>=3.1,<4
email-reply-parser
django-markdown-deux
xhtml2pdf
python-magic
XlsxWriter==0.9.6